+++ src/api/consumer/lib/migrate.ts
@@ -15,12 +15,12 @@ export default (async function migrate(
   scopePath: string,
   verbose: boolean
 ): Promise<MigrationResult | null | undefined> {
-  logger.debug('migrate.migrate, starting migration process');
+  logger.silly('migrate.migrate, starting migration process');
   if (verbose) console.log('starting migration process'); // eslint-disable-line no-console
   let scope;
   // If a scope path provided we will run the migrate only for the scope
   if (scopePath) {
-    logger.debug(`migrate.migrate, running migration process for scope in path ${scopePath}`);
+    logger.silly(`migrate.migrate, running migration process for scope in path ${scopePath}`);
     if (verbose) console.log(`running migration process for scope in path ${scopePath}`); // eslint-disable-line no-console
     scope = await loadScope(scopePath);
     return scope.migrate(verbose);
@@ -34,7 +34,7 @@ export default (async function migrate(
   await consumer.migrate(verbose);
   // const consumerMigrationResult = await consumer.migrate(verbose);
   // if (!consumerMigrationResult)
-  logger.debug('migrate.migrate, running migration process for scope in consumer');
+  logger.silly('migrate.migrate, running migration process for scope in consumer');
   if (verbose) console.log('running migration process for scope in consumer'); // eslint-disable-line no-console
   return scope.migrate(verbose);
 });