+++ src/scope/models/model-component.ts
@@ -455,7 +455,7 @@ export default class Component extends BitObject {
   }
 
   validateBeforePersisting(componentStr: string): void {
-    logger.debug(`validating component object: ${this.hash().hash} ${this.id()}`);
+    logger.silly(`validating component object: ${this.hash().hash} ${this.id()}`);
     const component = Component.parse(componentStr);
     component.validate();
   }